requirements.txt
# FastAPI and web server
fastapi>=0.115.10
uvicorn>=0.33.0
pydantic>=2.10.6
pydantic-settings>=2.2.1
email-validator>=2.1.1

# OpenAI
openai>=1.65.3

# HTTP client
httpx>=0.27.0

# Configuration
pyyaml>=6.0.0

# Date handling
python-dateutil>=2.9.0
pytz>=2024.1  # For timezone handling

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
